% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rszvec.R
\name{rszvec_add}
\alias{rszvec_add}
\title{Add a single document to a collection}
\usage{
rszvec_add(col, id, vector)
}
\arguments{
\item{col}{Collection handle returned by [rszvec_open()].}

\item{id}{Character string identifier for the document.}

\item{vector}{Numeric vector embedding.}
}
\value{
`col`, invisibly.
}
\description{
Add a single document to a collection
}
