% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rszvec.R
\name{rszvec_delete}
\alias{rszvec_delete}
\title{Delete a document from a collection}
\usage{
rszvec_delete(col, id)
}
\arguments{
\item{col}{Collection handle returned by [rszvec_open()].}

\item{id}{Character string identifier of the document to delete.}
}
\value{
`col`, invisibly.
}
\description{
Delete a document from a collection
}
